Sort, reorganise and refactor Discord webhook functionality
